John conroy/move provenance (#1240)
* Move visualization

* Update imports

* Add changelog file

* Move provenance components into own dir

* Add changelog file

* Update paths

* Move provenance inside detail

* Update changelog files

* Move visualization into detail
