+++ context/app/static/js/components/Detail/provenance/ProvAnalysisDetails/ProvAnalysisDetails.spec.js
