+++ context/app/static/js/components/Detail/provenance/ProvGraph/index.js
