+++ context/app/static/js/components/Detail/provenance/ProvTableDerivedLink/style.js
