+++ context/app/static/js/components/Detail/provenance/ProvTabs/style.js
