+++ context/app/static/js/components/Detail/provenance/ProvVis/__tests__/fixtures/complex-cwl.json
