+++ context/app/static/js/components/Detail/provenance/ProvVis/__tests__/fixtures/real-cwl.json
