+++ context/app/static/js/components/Detail/provenance/ProvVis/__tests__/fixtures/simple-cwl.json
