+++ context/app/static/js/components/Detail/visualization/Visualization/index.js
