+++ context/app/static/js/components/Detail/visualization/VisualizationWrapper/style.js
