+++ context/app/static/js/pages/Dataset/Dataset.jsx
@@ -3,12 +3,12 @@ import Typography from '@material-ui/core/Typography';
 
 import { LightBlueLink } from 'js/shared-styles/Links';
 import Files from 'js/components/files/Files';
-import ProvSection from 'js/components/Detail/ProvSection';
+import ProvSection from 'js/components/Detail/provenance/ProvSection';
 import Summary from 'js/components/Detail/Summary';
 import Attribution from 'js/components/Detail/Attribution';
 import Protocol from 'js/components/Detail/Protocol';
 import MetadataTable from 'js/components/Detail/MetadataTable';
-import VisualizationWrapper from 'js/components/Detail/VisualizationWrapper';
+import VisualizationWrapper from 'js/components/Detail/visualization/VisualizationWrapper';
 import DetailLayout from 'js/components/Detail/DetailLayout';
 import SummaryItem from 'js/components/Detail/SummaryItem';
 import useSendUUIDEvent from 'js/components/Detail/useSendUUIDEvent';
