+++ context/app/static/js/pages/Sample/Sample.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Typography from '@material-ui/core/Typography';
 
-import ProvSection from 'js/components/Detail/ProvSection';
+import ProvSection from 'js/components/Detail/provenance/ProvSection';
 import Summary from 'js/components/Detail/Summary';
 import Attribution from 'js/components/Detail/Attribution';
 import Protocol from 'js/components/Detail/Protocol';
